example/07_logging/example_logging_03/requirements.txt
Flask==1.1.1
google-cloud-logging==1.12.1
